Vendor Entity changes

+++ src/main/java/com/marinamooringmanagement/service/impl/VendorServiceImpl.java
@@ -1,20 +1,33 @@
 package com.marinamooringmanagement.service.impl;
 
+import com.marinamooringmanagement.constants.AppConstants;
 import com.marinamooringmanagement.exception.DBOperationException;
 import com.marinamooringmanagement.exception.ResourceNotFoundException;
+import com.marinamooringmanagement.mapper.CountryMapper;
+import com.marinamooringmanagement.mapper.StateMapper;
 import com.marinamooringmanagement.mapper.VendorMapper;
+import com.marinamooringmanagement.model.entity.Country;
+import com.marinamooringmanagement.model.entity.State;
+import com.marinamooringmanagement.model.entity.User;
 import com.marinamooringmanagement.model.entity.Vendor;
 import com.marinamooringmanagement.model.request.BaseSearchRequest;
 import com.marinamooringmanagement.model.request.VendorRequestDto;
 import com.marinamooringmanagement.model.response.BasicRestResponse;
+import com.marinamooringmanagement.model.response.CountryResponseDto;
+import com.marinamooringmanagement.model.response.StateResponseDto;
 import com.marinamooringmanagement.model.response.VendorResponseDto;
+import com.marinamooringmanagement.repositories.CountryRepository;
+import com.marinamooringmanagement.repositories.StateRepository;
+import com.marinamooringmanagement.repositories.UserRepository;
 import com.marinamooringmanagement.repositories.VendorRepository;
+import com.marinamooringmanagement.security.config.LoggedInUserUtil;
 import com.marinamooringmanagement.service.VendorService;
 import com.marinamooringmanagement.utils.SortUtils;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
+import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -51,6 +64,24 @@ public class VendorServiceImpl implements VendorService {
     @Autowired
     private SortUtils sortUtils;
 
+    @Autowired
+    private StateRepository stateRepository;
+
+    @Autowired
+    private CountryRepository countryRepository;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private LoggedInUserUtil loggedInUserUtil;
+
+    @Autowired
+    private StateMapper stateMapper;
+
+    @Autowired
+    private CountryMapper countryMapper;
+
     /**
      * Fetches a list of vendors based on the provided search request parameters and search text.
      *
@@ -59,11 +90,16 @@ public class VendorServiceImpl implements VendorService {
      * @return a BasicRestResponse containing the results of the vendor search.
      */
     @Override
-    public BasicRestResponse fetchVendors(final BaseSearchRequest baseSearchRequest, final String searchText) {
+    public BasicRestResponse fetchVendors(final BaseSearchRequest baseSearchRequest, final String searchText, final HttpServletRequest request) {
         final BasicRestResponse response = BasicRestResponse.builder().build();
         response.setTime(new Timestamp(System.currentTimeMillis()));
         try {
             logger.info("API called to fetch all the vendors from the database");
+
+            final String loggedInUserRole = loggedInUserUtil.getLoggedInUserRole();
+            final Integer loggedInUserId = loggedInUserUtil.getLoggedInUserID();
+
+            final Integer customerOwnerId = request.getIntHeader("CUSTOMER_OWNER_ID");
 
             Specification<Vendor> spec = new Specification<Vendor>() {
                 @Override
@@ -77,6 +113,18 @@ public class VendorServiceImpl implements VendorService {
                                 criteriaBuilder.like(vendor.get("website"), "%" + searchText + "%")
                         ));
                     }
+
+                    if (loggedInUserRole.equals(AppConstants.Role.ADMINISTRATOR)) {
+                        if(customerOwnerId == -1) throw new RuntimeException("Please select a customer owner");
+                        predicates.add(criteriaBuilder.and(criteriaBuilder.equal(vendor.join("user").get("id"), customerOwnerId)));
+                    } else if (loggedInUserRole.equals(AppConstants.Role.CUSTOMER_OWNER)) {
+                        if (customerOwnerId != -1 && !customerOwnerId.equals(loggedInUserId))
+                            throw new RuntimeException("Not authorized to perform operations on customer with different customer owner id");
+                        predicates.add(criteriaBuilder.and(criteriaBuilder.equal(vendor.join("user").get("id"), loggedInUserId)));
+                    } else {
+                        throw new RuntimeException("Not Authorized");
+                    }
+
                     return criteriaBuilder.and(predicates.toArray(new Predicate[predicates.size()]));
                 }
             };
@@ -92,7 +140,15 @@ public class VendorServiceImpl implements VendorService {
             final List<VendorResponseDto> vendorResponseDtoList = vendorList
                     .getContent()
                     .stream()
-                    .map(vendor -> vendorMapper.mapToVendorResponseDto(VendorResponseDto.builder().build(), vendor))
+                    .map(vendor -> {
+                        VendorResponseDto vendorResponseDto = vendorMapper.mapToVendorResponseDto(VendorResponseDto.builder().build(), vendor);
+                        if(null != vendor.getUser()) vendorResponseDto.setUserId(vendor.getUser().getId());
+                        if (null != vendor.getState()) vendorResponseDto.setStateResponseDto(stateMapper.mapToStateResponseDto(StateResponseDto.builder().build(), vendor.getState()));
+                        if (null != vendor.getCountry()) vendorResponseDto.setCountryResponseDto(countryMapper.mapToCountryResponseDto(CountryResponseDto.builder().build(), vendor.getCountry()));
+                        if (null != vendor.getRemitState()) vendorResponseDto.setRemitStateResponseDto(stateMapper.mapToStateResponseDto(StateResponseDto.builder().build(), vendor.getRemitState()));
+                        if (null != vendor.getRemitCountry()) vendorResponseDto.setRemitCountryResponseDto(countryMapper.mapToCountryResponseDto(CountryResponseDto.builder().build(), vendor.getRemitCountry()));
+                        return vendorResponseDto;
+                    })
                     .collect(Collectors.toList());
 
             response.setMessage("List of vendors in the database");
@@ -100,7 +156,7 @@ public class VendorServiceImpl implements VendorService {
             response.setStatus(HttpStatus.OK.value());
         } catch (Exception e) {
             logger.error("Error occurred while fetching all the vendors from the database: {}", e.getLocalizedMessage());
-            response.setMessage("Error occurred while fetching list of vendors from the database");
+            response.setMessage(e.getLocalizedMessage());
             response.setContent(e.getMessage());
             response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
         }
@@ -113,19 +169,19 @@ public class VendorServiceImpl implements VendorService {
      * @param requestDto the vendor request DTO
      */
     @Override
-    public BasicRestResponse saveVendor(final VendorRequestDto requestDto) {
+    public BasicRestResponse saveVendor(final VendorRequestDto requestDto, final HttpServletRequest request) {
         final BasicRestResponse response = BasicRestResponse.builder().build();
         response.setTime(new Timestamp(System.currentTimeMillis()));
         try {
             logger.info("API called to save the vendor in the database");
             final Vendor vendor = Vendor.builder().build();
-            performSave(requestDto, vendor, null);
+            performSave(requestDto, vendor, null, request);
             response.setMessage("Vendor Saved Successfully");
             response.setStatus(HttpStatus.CREATED.value());
         } catch (Exception e) {
             logger.error("Error occurred while saving the vendor in the database {}", e.getLocalizedMessage());
             response.setMessage("Error occurred while saving the vendor in the database");
-            response.setContent(e.getMessage());
+            response.setContent(e.getLocalizedMessage());
             response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
         }
         return response;
@@ -138,18 +194,39 @@ public class VendorServiceImpl implements VendorService {
      * @return a message indicating the deletion status
      */
     @Override
-    public BasicRestResponse deleteVendor(final Integer vendorId) {
+    public BasicRestResponse deleteVendor(final Integer vendorId, final HttpServletRequest request) {
         final BasicRestResponse response = BasicRestResponse.builder().build();
         response.setTime(new Timestamp(System.currentTimeMillis()));
         try {
             logger.info("API called to delete the vendor from the database");
+
+            final String loggedInUserRole = loggedInUserUtil.getLoggedInUserRole();
+            final Integer loggedInUserID = loggedInUserUtil.getLoggedInUserID();
+
+            final Integer customerOwnerId = request.getIntHeader("CUSTOMER_OWNER_ID");
+
+            final Optional<Vendor> optionalVendor = vendorRepository.findById(vendorId);
+            if(optionalVendor.isEmpty()) throw new ResourceNotFoundException(String.format("No vendor found with the given id: %1$s", vendorId));
+            final Vendor vendor = optionalVendor.get();
+
+            if(loggedInUserRole.equals(AppConstants.Role.ADMINISTRATOR)) {
+                if(customerOwnerId == -1) throw new RuntimeException("Please select a customer owner");
+                else if(!customerOwnerId.equals(vendor.getUser().getId())) throw new RuntimeException("Cannot perform operations on vendor with different customer owner id");
+            } else if (loggedInUserRole.equals(AppConstants.Role.CUSTOMER_OWNER)) {
+                if(customerOwnerId != -1 && !loggedInUserID.equals(customerOwnerId)) throw new RuntimeException("Cannot perform operations on vendor with different customer owner id");
+                if (!vendor.getUser().getId().equals(loggedInUserID))
+                    throw new RuntimeException("Not authorized to perform operations on vendor with different customer owner Id");
+            } else{
+                throw new RuntimeException("Not Authorized");
+            }
+
             vendorRepository.deleteById(vendorId);
             final String message = vendorRepository.findById(vendorId).isPresent() ? String.format("Vendor with the id %1$s failed to get deleted", vendorId) : String.format("Vendor with the id %1$s is deleted successfully", vendorId);
             response.setMessage(message);
             response.setStatus(HttpStatus.OK.value());
         } catch (Exception e) {
             logger.error("Error occurred while deleting the vendor from the database {}", e.getLocalizedMessage());
-            response.setMessage("Error occurred while deleting the vendor from the database");
+            response.setMessage(e.getLocalizedMessage());
             response.setContent(e.getMessage());
             response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
         }
@@ -160,29 +237,29 @@ public class VendorServiceImpl implements VendorService {
      * Updates a vendor based on the provided request DTO and vendor ID.
      *
      * @param requestDto the vendor request DTO
-     * @param vendorId   the vendor ID
+     * @param id   the vendor ID
      */
     @Override
-    public BasicRestResponse updateVendor(final VendorRequestDto requestDto, final Integer vendorId) {
+    public BasicRestResponse updateVendor(final VendorRequestDto requestDto, final Integer id, final HttpServletRequest request) {
         final BasicRestResponse response = BasicRestResponse.builder().build();
         response.setTime(new Timestamp(System.currentTimeMillis()));
         try {
             logger.info("API called to update the vendor in the database");
-            if(null == vendorId) {
-                throw new RuntimeException("vendorId cannot be null during update command");
+            if(null == id) {
+                throw new RuntimeException("Id cannot be null during update command");
             } else {
-                Optional<Vendor> optionalVendor = vendorRepository.findById(vendorId);
+                Optional<Vendor> optionalVendor = vendorRepository.findById(id);
                 if(optionalVendor.isEmpty()) {
                     throw new ResourceNotFoundException("No vendor exists with the given vendor ID");
                 }
                 final Vendor vendor = optionalVendor.get();
-                performSave(requestDto, vendor, vendorId);
+                performSave(requestDto, vendor, id, request);
             }
             response.setMessage("Vendor updated successfully");
             response.setStatus(HttpStatus.OK.value());
         } catch (Exception e) {
             logger.error("Error occurred while updating the vendor in the database {}", e.getLocalizedMessage());
-            response.setMessage("Error occurred while updating the vendor in the database");
+            response.setMessage(e.getLocalizedMessage());
             response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
         }
         return response;
@@ -191,25 +268,90 @@ public class VendorServiceImpl implements VendorService {
     /**
      * Performs the actual saving of a vendor entity based on the request DTO and vendor object.
      *
-     * @param requestDto the vendor request DTO
+     * @param vendorRequestDto the vendor request DTO
      * @param vendor     the vendor object to be saved or updated
-     * @param vendorId   the vendor ID (null for new vendors)
+     * @param id   the vendor ID (null for new vendors)
      */
-    public Vendor performSave(final VendorRequestDto requestDto, final Vendor vendor, final Integer vendorId) {
+    public Vendor performSave(final VendorRequestDto vendorRequestDto, final Vendor vendor, final Integer id, final HttpServletRequest request) {
         final Vendor savedVendor;
         try {
+            final String loggedInUserRole = loggedInUserUtil.getLoggedInUserRole();
+            final Integer loggedInUserId = loggedInUserUtil.getLoggedInUserID();
             logger.info("performSave() function called");
-            if(null == vendorId) {
+            if(null == id) {
                 vendor.setCreationDate(new Date(System.currentTimeMillis()));
                 vendor.setLastModifiedDate(new Date(System.currentTimeMillis()));
             } else {
                 vendor.setLastModifiedDate(new Date(System.currentTimeMillis()));
             }
-            vendorMapper.mapToVendor(vendor, requestDto);
+
+            Integer customerOwnerId = request.getIntHeader("CUSTOMER_OWNER_ID");
+
+            User user = null;
+            if (loggedInUserRole.equals(AppConstants.Role.ADMINISTRATOR)) {
+                if(customerOwnerId == -1) throw new RuntimeException("Please select a customer owner");
+                Optional<User> optionalUser = userRepository.findById(customerOwnerId);
+                if (optionalUser.isEmpty())
+                    throw new ResourceNotFoundException(String.format("No user found with the given id: %1$s", customerOwnerId));
+                if (!optionalUser.get().getRole().getName().equals(AppConstants.Role.CUSTOMER_OWNER))
+                    throw new RuntimeException(String.format("User with the given id: %1$s is not of Customer Owner role.", customerOwnerId));
+                user = optionalUser.get();
+            } else if (loggedInUserRole.equals(AppConstants.Role.CUSTOMER_OWNER)) {
+                if (customerOwnerId != -1 && !loggedInUserId.equals(customerOwnerId))
+                    throw new RuntimeException("Cannot do operations on vendor with different customer owner Id");
+                customerOwnerId = loggedInUserId;
+                Optional<User> optionalUser = userRepository.findById(loggedInUserId);
+                if (optionalUser.isEmpty())
+                    throw new ResourceNotFoundException(String.format("No user found with the given id: %1$s", customerOwnerId));
+                user = optionalUser.get();
+            } else {
+                throw new RuntimeException("Not Authorized");
+            }
+
+            vendor.setUser(user);
+
+            vendorMapper.mapToVendor(vendor, vendorRequestDto);
+
+            if (null != vendorRequestDto.getStateId()) {
+                final Optional<State> optionalState = stateRepository.findById(vendorRequestDto.getStateId());
+                if (optionalState.isEmpty())
+                    throw new ResourceNotFoundException(String.format("No state found with the given Id: %1$s", vendorRequestDto.getStateId()));
+                vendor.setState(optionalState.get());
+            } else {
+                if (null == id) throw new RuntimeException("State cannot be null.");
+            }
+
+            if (null != vendorRequestDto.getCountryId()) {
+                final Optional<Country> optionalCountry = countryRepository.findById(vendorRequestDto.getCountryId());
+                if (optionalCountry.isEmpty())
+                    throw new ResourceNotFoundException(String.format("No country found with the given Id: %1$s", vendorRequestDto.getCountryId()));
+                vendor.setCountry(optionalCountry.get());
+            } else {
+                if (null == id) throw new RuntimeException("Country cannot be null.");
+            }
+
+            if (null != vendorRequestDto.getRemitStateId()) {
+                final Optional<State> optionalState = stateRepository.findById(vendorRequestDto.getRemitStateId());
+                if (optionalState.isEmpty())
+                    throw new ResourceNotFoundException(String.format("No state found with the given Id: %1$s", vendorRequestDto.getRemitStateId()));
+                vendor.setRemitState(optionalState.get());
+            } else {
+                if (null == id) throw new RuntimeException("State cannot be null.");
+            }
+
+            if (null != vendorRequestDto.getRemitCountryId()) {
+                final Optional<Country> optionalCountry = countryRepository.findById(vendorRequestDto.getRemitCountryId());
+                if (optionalCountry.isEmpty())
+                    throw new ResourceNotFoundException(String.format("No country found with the given Id: %1$s", vendorRequestDto.getRemitCountryId()));
+                vendor.setRemitCountry(optionalCountry.get());
+            } else {
+                if (null == id) throw new RuntimeException("Country cannot be null.");
+            }
+
             savedVendor = vendorRepository.save(vendor);
         } catch (Exception e) {
             logger.error("Error occurred during performSave() operation {}", e.getLocalizedMessage());
-            throw new DBOperationException("Error occurred during performSave() operation", e);
+            throw e;
         }
         return savedVendor;
     }
